change EPP on basis of charging state and battery capacity, closes #1

+++ service-enable.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
 
 SERVICE_FILE="default-energy-performance-preference.service"
+CPU_SCRIPT="set_cpu_performance.sh"
+TIMER="default-energy-performance-preference.timer"
 SYSTEMD_DIR="/etc/systemd/system"
 
 if [ ! -f "$SERVICE_FILE" ]; then
@@ -8,9 +10,23 @@ if [ ! -f "$SERVICE_FILE" ]; then
   exit 1
 fi
 
+if [ ! -f "$CPU_SCRIPT" ]; then
+  echo "Error: $CPU_SCRIPT not found in the current directory."
+  exit 1
+fi
+
+if [ ! -f "$TIMER" ]; then
+  echo "Error: $TIMER not found in the current directory."
+  exit 1
+fi
+
 sudo cp "$SERVICE_FILE" "$SYSTEMD_DIR"
-pkexec systemctl daemon-reload
-pkexec systemctl enable "$SERVICE_FILE"
-pkexec systemctl start "$SERVICE_FILE"
+sudo cp "$CPU_SCRIPT" "/usr/local/bin"
+sudo cp "$TIMER" "$SYSTEMD_DIR"
+sudo systemctl daemon-reload
+sudo systemctl enable "$TIMER"
+sudo systemctl start "$TIMER"
+sudo systemctl enable "$SERVICE_FILE"
+sudo systemctl start "$SERVICE_FILE"
 
 echo "Service has been enabled and started."
